Optimize Python propensity score matcher
Test various optimizations for the Python propensity score matcher
to see if we can reduce the runtime.

Conclusion:

Optimizations barely reduce the runtime. It will be extremely difficult
to make a pure Python program faster, and we are better off
rewriting the function in C++ and optimizing the C++ program.