Change tree index values
Change the search tree from using indicies 0 to N-1 to indicies
1 to N.

This change makes it easier to delete the root node. The value at
position zero now stores the position of the root node.

Building the tree was moved to a function, and the search and tree
construction functions were re-written to use the new indicies.